* master: (42 commits)
  remove unnecessary checks from "fire" events. Furthermore, 'new_record?' does not work after 'create' ("create" is a call to 'repository.~save~' sets @new_record false!)
  fix activity logs - checks are necessary ("#save just delegates to #update or #create internally depending on new_record?"*)
  fix: set_published_permalink needs to work for a newly created article as well
  Fixing more DM 0.9 fun issues with include order
  fixing the Merb 0.9.3 cache issue at long last within Feather, as trying to get people to use only Merb 0.9.2 is becoming difficult - this can be removed once Merb 0.9.4 is released (the cache issue is fixed within edge)
  added a db task to auto migrate all core classes for feather (for initial database setup)
  Changed destroy! to destroy to be compatible with DM 0.9.1
  Fixing an issue with datamapper 0.9.1 where before :save throws an super: no superclass method `save' - (NoMethodError) error. Changed all before :saves to before :create and :update respectively, except for the article save_events, and moved include DataMapper::Resource to the bottom of the include list. I think that was the real fix.
  few changes necessary for DataMapper 0.9, few other fixes
  few changes necessary for DM 0.9 integration
  Taking out merb edge rack adapter change for now
  ok, sorry, folks! revert 99% of the last patch (we're moving forward with specifying the full url for a plugin). In the meantime, preserve the "mkdir fix" for app/plugins
  fix plugin installation problems no. 1 (***Also fixes "url to key in for plugin installation" problem (*see below if you need more info))
  fixed plugin installation for those with gems, also added the missing "link_to_author" helper method
  moving require statements needed in plugin model to init
  match not needed here, just try everything.  This fixes permalinks not working with Merb edge
  Updated rack adapter to lastest Merb edge
  Finished plugin packages
  Fix for PluginSetting#write
  Fix target path for plugins with version numbers
  ...

+++ app/models/plugin.rb
@@ -1,18 +1,23 @@
-class Plugin < DataMapper::Base
-  property :url, :string, :length => 255
-  property :path, :string, :length => 255
-  property :name, :string
-  property :author, :string
-  property :version, :string
-  property :homepage, :string, :length => 255
-  property :about, :string, :length => 255
-  property :active, :boolean
+class Plugin
+  include DataMapper::Resource
 
-  before_create :download
-  after_create :install
-  after_create :set_create_activity
-  after_update :set_update_activity
-  after_destroy :remove
+  property :id, Integer, :key => true, :serial => true
+  property :url, String, :length => 255
+  property :path, String, :length => 255
+  property :name, String, :length => 255
+  property :version, String, :length => 255
+  property :author_name, String, :length => 255
+  property :author_email, String, :length => 255
+  property :author_homepage, String, :length => 255
+  property :homepage, String, :length => 255
+  property :about, String, :length => 255
+  property :active, Boolean
+
+  before :save, :download
+  after :save, :install
+  after :save, :set_create_activity
+  after :save, :set_update_activity
+  after :destroy, :remove
 
   class << self
     @@loaded = []
@@ -21,51 +26,70 @@ class Plugin < DataMapper::Base
   ##
   # This grabs the plugin using its url, unpacks it, and loads the metadata for it
   def download
-    # Load the manifest yaml
-    manifest = YAML::load(Net::HTTP.get(URI.parse(url + "/manifest.yml")))
-    # Grab metadata from manifest
-    self.name = manifest["plugin"]["name"]
-    self.author = manifest["plugin"]["author"]
-    self.version = manifest["plugin"]["version"]
-    self.homepage = manifest["plugin"]["homepage"]
-    self.about = manifest["plugin"]["about"]
-    # Build the path
-    self.path = File.join(File.join(File.join(Merb.root, "app"), "plugins"), URI.parse(url).path.split("/").last.split(".").first)
-    # Remove any existing plugin at the path
-    FileUtils.rm_rf(self.path)
-    # Download all of the plugin contents
-    recurse(manifest["plugin"]["contents"])
-    # Unpack any gems downloaded
-    unpack_gems(manifest["plugin"]["contents"]["gems"]["."]) unless manifest["plugin"]["contents"]["gems"].nil?
+    if new_record?
+      # Load the manifest yaml
+      manifest = YAML::load(Net::HTTP.get(URI.parse(url)))
+      # Grab metadata from manifest
+      self.name = manifest["name"]
+      self.author = manifest["author"]
+      self.version = manifest["version"]
+      self.homepage = manifest["homepage"]
+      self.about = manifest["about"]
+      # Build the path
+      self.path = File.join(Merb.root, "app", "plugins", self.name)
+      # Remove any existing plugin at the path
+      FileUtils.rm_rf(self.path)
+      FileUtils.mkdir_p(self.path)
+      # Download the package and untgz
+      package_url = File.join(url.split('/').slice(0..-2).join('/'), manifest["package"])
+      package = Net::HTTP.get(URI.parse(package_url))
+      Archive::Tar::Minitar.unpack(Zlib::GzipReader.new(StringIO.new(package)), self.path)
+      # Unpack any gems downloaded
+      unpack_gems(Dir.glob(File.join(self.path, "gems", "*.gem")).collect { |p| p.split("/").last })
+    end
   end
 
   ##
   # This loads and installs the plugin
   def install
-    # Load the plugin
-    self.load
-    # Also, if there is an "install.rb" script present, run that to setup anything the plugin needs (database tables etc)
-    require File.join(self.path, "install.rb") if File.exists?(File.join(self.path, "install.rb"))
-  rescue Exception => err
-    # Catch the error, delete the plugin, and raise an error again
-    self.destroy!
-    raise "Error installing plugin: #{err.message}!"
+    if new_record?
+      # Ensure we don't have any errors while saving already
+      if self.errors.empty?
+        begin
+          # Load the plugin
+          self.load
+          # Also, if there is an "install.rb" script present, run that to setup anything the plugin needs (database tables etc)
+          require File.join(self.path, "install.rb") if File.exists?(File.join(self.path, "install.rb"))
+        rescue Exception => err
+          # If we have an issue installing, lets destroy the plugin to rollback, and put an error on the object so it displays on the form
+          self.destroy
+          self.errors.add "Error installing plugin: #{err.message}"
+        end
+      else
+        # If we do, destroy the plugin to rollback, and then the form will display the errors
+        self.destroy
+      end
+    end
   end
 
   ##
   # This adds the activity to show a plugin has been installed
   def set_create_activity
-    a = Activity.new
-    a.message = "Plugin \"#{self.name}\" installed"
-    a.save
+    if new_record?
+      a = Activity.new
+      a.message = "Plugin \"#{self.name}\" installed"
+      a.save
+    end
   end
 
   ##
   # This adds the activity to show a plugin has been updated
   def set_update_activity
-    a = Activity.new
-    a.message = "Plugin \"#{self.name}\" #{self.active ? 'activated' : 'de-activated'}"
-    a.save
+    unless new_record?
+      a = Activity.new
+      a.message = "Plugin \"#{self.name}\" #{self.active ? 'activated' : 'de-activated'}"
+      a.save
+    end
   end
 
   ##
@@ -81,18 +105,36 @@ class Plugin < DataMapper::Base
   ##
   # This loads the plugin, first loading any gems it may have
   def load
-    # Setup the Gem path to the plugin
-    Gem.use_paths(Gem.dir, ((Gem.path - [Gem.dir]) + [self.path]))
-    # Load the plugin init script
-    require File.join(self.path, "init.rb")
-    # Add the plugin to the array of loaded plugins
-    @@loaded << self.name
+    # Plugin dependencies let you load a plugin before this one,
+    # so we don't want to load that sucker twice, now do we?
+    unless loaded?
+      # Setup the Gem path to the plugin
+      Gem.use_paths(Gem.dir, ((Gem.path - [Gem.dir]) + [self.path]))
+      # Load the plugin init script
+      require File.join(self.path, "init.rb")
+      # Add the plugin to the array of loaded plugins
+      @@loaded << self.name
+    end
   end
 
   ##
   # This returns true if the plugin has been loaded, false otherwise
   def loaded?
     @@loaded.include?(self.name)
+  end
+  
+  def author
+    {
+      'name' => author_name,
+      'email' => author_email,
+      'homepage' => author_homepage
+    }
+  end
+  
+  def author=(author)
+    %W{name email homepage}.each do |k|
+      self.send("author_#{k}=".to_sym, author[k])
+    end
   end
 
   private
@@ -122,9 +164,7 @@ class Plugin < DataMapper::Base
     def unpack_gems(gems)
       gems.each do |gem|
         # Unpack the gem
-        `gem unpack #{File.join(File.join(self.path, "gems"), gem)}`
-        # We can't seem to use --target on the gem command above to actually specify the output folder - so it's in Merb.root; lets move it
-        FileUtils.mv gem.gsub(".gem", ""), File.join(File.join(self.path, "gems"), gem.gsub(".gem", ""))
+        `cd #{File.join(self.path, "gems")}; gem unpack #{File.join(File.join(self.path, "gems"), gem)}`
       end
     end
 end
